🔵refactor: Gamer 추상클래스 제거
