+++ src/main/java/domain/GameManager.java
@@ -2,6 +2,7 @@ package domain;
 
 import domain.gamer.Dealer;
 import domain.gamer.Player;
+
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
