+++ src/main/java/domain/gamer/Player.java
@@ -1,31 +1,55 @@
 package domain.gamer;
 
+import domain.GameResult;
+import domain.card.Card;
 import domain.card.CardGenerator;
 import domain.card.CardGroup;
-import java.util.Objects;
+import java.util.List;
 
-public class Player extends Gamer {
+public class Player {
+    public static final int LIMIT = 21;
+
+    protected final CardGroup cardGroup;
+    protected final CardGenerator cardGenerator;
     private final String name;
 
+
     public Player(String name, CardGroup cardGroup, CardGenerator cardGenerator) {
-        super(cardGroup, cardGenerator);
         this.name = name;
+        this.cardGroup = cardGroup;
+        this.cardGenerator = cardGenerator;
+    }
+
+    public void receiveCard(){
+        cardGroup.addCard(cardGenerator.generate());
+    }
+
+    public void receiveCard(int count) {
+        while (count-- > 0) {
+            receiveCard();
+        }
+    }
+
+    public boolean isBust() {
+        return this.cardGroup.calculateScore(LIMIT) > LIMIT;
+    }
+
+    public GameResult calculateGameResult(final int compareScore) {
+        if (isBust()) {
+            return GameResult.LOSE;
+        }
+        return GameResult.findByScores(cardGroup.calculateScore(LIMIT), compareScore);
+    }
+
+    public int calculateScore() {
+        return cardGroup.calculateScore(LIMIT);
+    }
+
+    public List<Card> getCards() {
+        return cardGroup.getCards();
     }
 
     public String getName() {
         return name;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (!(o instanceof final Player player)) {
-            return false;
-        }
-        return Objects.equals(getName(), player.getName());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getName());
     }
 }
